replace tests; iss #3
